+++ lib/_stream_duplex.js
@@ -29,15 +29,33 @@
 module.exports = Duplex;
 
 const util = require('util');
-const DuplexBase = require('internal/streams/duplex_base');
+const Readable = require('_stream_readable');
+const Writable = require('_stream_writable');
 
-util.inherits(Duplex, DuplexBase);
+util.inherits(Duplex, Readable);
+
+{
+  // Allow the keys array to be GC'ed.
+  const keys = Object.keys(Writable.prototype);
+  for (var v = 0; v < keys.length; v++) {
+    const method = keys[v];
+    if (!Duplex.prototype[method])
+      Duplex.prototype[method] = Writable.prototype[method];
+  }
+}
 
 function Duplex(options) {
   if (!(this instanceof Duplex))
     return new Duplex(options);
 
-  DuplexBase.call(this, options);
+  Readable.call(this, options);
+  Writable.call(this, options);
+
+  if (options && options.readable === false)
+    this.readable = false;
+
+  if (options && options.writable === false)
+    this.writable = false;
 
   this.allowHalfOpen = true;
   if (options && options.allowHalfOpen === false) {
