+++ lib/net.js
@@ -75,7 +75,6 @@ const {
   ERR_SOCKET_BAD_PORT,
   ERR_SOCKET_CLOSED
 } = errors.codes;
-const DuplexBase = require('internal/streams/duplex_base');
 const dns = require('dns');
 
 const kLastWriteQueueSize = Symbol('lastWriteQueueSize');
@@ -242,19 +241,19 @@ function Socket(options) {
 
   if (typeof options === 'number')
     options = { fd: options }; // Legacy interface.
-  else if (options === undefined)
-    options = {};
   else
     options = util._extend({}, options);
 
+  const allowHalfOpen = options.allowHalfOpen;
+
+  // Prevent the "no-half-open enforcer" from being inherited from `Duplex`.
+  options.allowHalfOpen = true;
   // For backwards compat do not emit close on destroy.
   options.emitClose = false;
+  stream.Duplex.call(this, options);
 
-  // `DuplexBase` is just a slimmed down constructor for `Duplex` which allow
-  // us to not inherit the "no-half-open enforcer" as there is already one in
-  // place. Instances of `Socket` are still instances of `Duplex`, that is,
-  // `socket instanceof Duplex === true`.
-  DuplexBase.call(this, options);
+  // Default to *not* allowing half open sockets.
+  this.allowHalfOpen = Boolean(allowHalfOpen);
 
   if (options.handle) {
     this._handle = options.handle; // private
@@ -299,9 +298,6 @@ function Socket(options) {
 
   // handle strings directly
   this._writableState.decodeStrings = false;
-
-  // default to *not* allowing half open sockets
-  this.allowHalfOpen = options.allowHalfOpen || false;
 
   // if we have a handle, then start the flow of data into the
   // buffer.  if not, then this will happen when we connect
